+++ aptihramy-frontend/src/pages/HomePage.vue
@@ -24,14 +24,14 @@
 import DisplayPeople from '@/components/DisplayPeople.vue';
 import { ref, computed, watch, onMounted } from 'vue';
 import Filter from '@/components/Filter.vue';
-import { FilterState, TrackerIDMemory } from "../types/types"
-import { FilterRequest, } from '@/types/api';
+import { FilterRequest, } from '@/types/api/api';
 import '../styles/main.css';
 import { fetchFilteredTrackers, UNAUTHORIZED } from '@/core/api';
 import { useErrorMessagesStore } from '@/core/stores/errorMessages';
 import { useTrackedFeaturesStore } from '@/core/stores/trackedFeatures';
 import { useTrackedYearsStore } from '@/core/stores/trackedYears';
 import { useFilterStore } from '@/core/stores/filterStore';
+import { FilterState, TrackerIDMemory } from '@/types';
 
 
 const errorMessageStore = useErrorMessagesStore()
