+++ aptihramy-frontend/src/types/api/api.ts
@@ -6,3 +6,4 @@ export * from './tracker';
 export * from './records';
 export * from './filters';
 export * from './chain';
+export * from './update';
