+++ aptihramy-frontend/src/types/api/tracker.ts
