+++ aptihramy-frontend/src/types/components/candidates.ts
@@ -1,0 +1,7 @@
+export interface CandidateRecordValues {
+    raw_value: string | number;
+    normalized_value: string | number | null;
+    memory: (string | number)[] | null;
+    distances: number[] | null;
+    score: number | null;
+}
