+++ aptihramy-frontend/src/types/components/frame.ts
@@ -1,0 +1,9 @@
+import { RecordValuesDiag } from "../api/api";
+
+export interface OneFrameInformationProps {
+    recordValuesDiagnostic: RecordValuesDiag | null;
+    frameIdx: number;
+    recordIdx: number;
+    memory: string[][] | null;
+    nbColumns: number;
+}
