+++ aptihramy-frontend/src/types/components/people.ts
@@ -1,0 +1,6 @@
+import { TrackerIDMemory } from "../base";
+
+export interface DisplayPeopleProps {
+    data: TrackerIDMemory;
+    isLoading: boolean;
+}
