+++ aptihramy-frontend/src/types/components/snackbar.ts
