+++ aptihramy-frontend/src/types/components/topbar.ts
@@ -1,0 +1,16 @@
+export interface TopBarProps {
+    resetZoom?: Function;
+    goToEditPage?: Function;
+    title: string;
+}
+
+export interface TopBarTrackingChainProps {
+    resetZoom: Function;
+    goToEditPage: Function;
+    title: string;
+}
+
+export interface TopBarEditPageProps {
+    save: Function;
+    title: string;
+}
