+++ aptihramy-frontend/src/types/components/tracking.ts
@@ -1,0 +1,5 @@
+import { trackerID } from "../api/api";
+
+export interface TrackinChainProps {
+    trackerID: trackerID;
+}
