+++ aptihramy-frontend/src/types/index.ts
@@ -1,0 +1,11 @@
+export * from "./base";
+export * from "./feature";
+export * from "./emits";
+export * from "./components/snackbar"
+export * from "./components/topbar";
+export * from "./components/filter";
+export * from "./components/edit";
+export * from "./components/people";
+export * from "./components/frame";
+export * from "./components/tracking";
+export * from "./components/candidates";
